Resolved naming conflict
Solidworks has a problem with different parts having the same name. This problem was brought because of the PCBs having the same parts name like BOARD.SLDPRT and R1.SLDPRT. A part could have been generated from the assembly but it was too memory and CPU hungry to be produced so all the part were renamed to have unique name thus preventing naming conflict.